Fix more MPS issues

+++ src/data/hf_dataset.py
@@ -42,7 +42,7 @@ class HFDataset(Dataset):
             v2.Resize(256),
             v2.CenterCrop(224),
             v2.ToImage(),
-            v2.ToDtype(torch.bfloat16, scale=True),
+            v2.ToDtype(torch.float32, scale=True),
             v2.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
         ])
